Batteries

Categorizing repair attempts on battery-powered devices recorded at community events.

batteries

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 995 rows and 12 columns.

Source

https://github.com/openrepair/data/tree/master/quests/batteries

Details

Dataset of battery-powered device repair records from the Open Repair Alliance in ORDS format style with added step_barrier values. Results of Open Repair Data quest BattCat.

id_ords

ID. Unique identifier from the partner organisation.

data_provider

Data provider. Option from ORDS codelist. Name of partner organisation.

country

Country. Three letter ISO code, e.g. “GBR”.

step_barrier_id

Fault type ID.

step_barrier

Fault type. Name of the step or barrier. Option from battery codelist.

product_category

Product category. Option from ORDS product category codelist

brand

Brand.

repair_status

Repair status option from the battery codelist.

year_of_manufacture

Year of manufacture.

event_date

Event date. The date of the repair event that the repair took place at.

language

Language of problem description. 2 char ISO language code.

problem

Problem description.