Mobiles

Result of two online quests where volunteer participants looked at about 1900 mobile phone repair attempts recorded at community events.

mobiles

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 1916 rows and 11 columns.

Source

https://github.com/openrepair/data/tree/master/quests/mobiles

Details

Dataset of printer repair records from the Open Repair Alliance in ORDS format style with added fault_type values. Results of Results of Open Repair Data quests MobiFix and MobiFix:ORA.

There are some differences in the structure of this data and any of the ORDS format versions (0.1 and 0.2). The records involved span different ORDS versions and a handful have not yet been exported.

The usual ORDS id values have been replaced by special mobifix-specific unique identifiers

The model field was removed from the ORDS v0.2 standard but was useful for these quests and so has been retained in this dataset

Age is an important metric in repair policy, therefore it was derived from and replaced the year_of_manufacture field

The field group_identifier is irrelevent in this analysis and thus has been omitted from the dataset

MobiFix was held in July 2020 using data from ORA partner The Restart Project. MobiFix:ORA was held in March 2021 using data from ORA partners anstiftung and Repair Café International.

id

ID.

data_provider

Data provider. Option from ORDS codelist. Name of partner organisation.

country

Country. Three letter ISO code, e.g. “GBR”.

fault_type

Fault type.

product_category

Product category. Option from ORDS product category codelist.

brand

Brand.

model

Product model.

repair_status

Result of the repair attempt.

age

Product age.

event_date

Event date. The date of the repair event that the repair took place at.

problem

Problem description.